Added artifacts to make it easier to post process KapeFiles.Targets
The Windows.KapeFiles.Targets artifact allows to collect many bulk
forensic artifacts like registry hives etc. People often use it to
collect offline collections for preservation of hosts.

Although best practice is to **also** collect parsing artifacts as the
same time, sometimes this is left out. In this case the user needs to
parse the collected raw files (for example collecting $MFT then
needing to apply `Windows.NTFS.MFT` to parse it). This is not always
easy to do.

This PR adds a notebook suggestion which applies remapping on the
KapeFiles collection in such as way that some regular artifacts
designed to run on the live system can work to some extent off the raw
collection. The additional parsing will be added to new collections in
the same (possibly virtual) client.

Also: Bugfix for zip_nocase accessor failed to work in some cases.

Since we now use remapping there is no need for the
Linux.KapeFiles.CollectFromDirectory (which was basically the same as
the Windows.KapeFiles.Targets but applies to a mounted
directory). That artifact is now removed.