Refactored timelines (#3699)
1. Timelines now specify 3 standard columns which are always present: a.
The timestamp which the timeline is keyed on. b. A message which is
shows to describe the event. c. A timestamp description which describes
the type of the event.
2. These can be any column in the added data set but on timeline
addition we can specify which column is which.
3. New UI for examining timeline events shows these three columns and
does not show any of the other events. This keeps the display cleaner
and more concise.
4. A user can click on any event row to see all the other columns.
5. Added a column selector: Users can add additional columns to the
concise table for review. This makes it easy to compare similar events
side by side.

Added the following VQL plugins:
- timeline_delete: Allows deleting an existing timeline.
- timelines: Lists the timelines in a notebook
- timeline_add: revamped to allow the additional column definitions.