Refactor client delete into client info service.
Also fix a bug where when clients were removed from the client info
the snapshot was not marked as dirty so did not cause a flush to
disk. This caused the client to be resurrected on next restart.

This PR also rebuilds the search index periodically to create a more
stable index when clients are quickly added and removed