Use HostID as basis for client id when importing offline collection
When importing an offline collection a new client is usually
created. If the same host gets collected again we use the hostname to
try to resolve the client id and add the collection to the new host
bit since the offline endpoint is not a real client it does not have a
permanent client id.

In this PR we store the host UUID (which most OSes provide) in the
offline collection itself. This should not be changing over the life
of the host so we can use it as the basis for the client id. This
ensures the same client is used for all collections from the same
host.