Client metadata fields can now be indexed and searched. (#3720)
A new config option `defaults.indexed_client_metadata` can now specify a
list of fields in the client metadata fields. When the admin sets these
fields, Velociraptor will add the fields to the client search index.
This means it is possible to now search on these custom fields in the
search bar.

For example, say we add the field "dept" to signify a client's
Department:

```yaml
defaults:
  indexed_client_metadata:
    - dept
```

Adding a client metadata field `dept` with value `engineering` will
allow us to now search for the client using the verb `dept:engineering`
in the GUI