Wrap tempfile usage to work around Windows/Golang bug
This is a workaround to the introduction of GetTempPath2 in
Windows. The Golang standard library switched to this API in recent Go
version https://github.com/golang/go/issues/56899

The GetTempPath2 API ignores the TEMP environment variable and
always returns C:\Windows\SystemTemp for system level programs. This
causes the os.TempDir() API to ignore our configuration of the
correct temp location.

Since practically, we can not allow list the SystemTemp directory in
e.g. EDR or other security products, this bug forces Velociraptor to
write temp files in this common area ignoring the configuration
file.

To work around this issue we wrap all calls to temporary files and
ignore the os.TempDir() API completely.