Allows arbitrary length files in filestore.
On Linux file names must be less then 255 bytes. The filestore
attempts to preserve and sanitize the original filename of uploaded
files, which means that in some cases the filestore filename can
easily exceed the hard limit imposed by the OS.

This PR switches to a hash based compression when the filename exceeds
250 chars. This replaces the filestore filename with a hash, and the
full component is stored in the datastore separately. This allows us
to store arbitrary length filenames in the filestore.