Usability enhancements for table GUI (#3792)
Allow the whole table, specific columns, or specific cells to be
compacted. This helps certain data which may be too large to fit in a
cell. Previously such data caused the cell to be expanded which made the
table more sparse and hard to see. This PR allows only certain cells to
be expanded as needed to keep the table compact taking less real estate.

Reworked the filter/sort buttons at the top of each column to appear
outside the header itself. Previously these buttons were part of the
header causing the header to be slightly wider than it needed to be-
making the table more sparse. This new CSS allows the buttons to extend
outside the column when they appear.

Changed the sense of drag and drop on column headers to insert the
dragged column before the dragged upon table - previously they would be
swapped but this is not intuitive.


![image](https://github.com/user-attachments/assets/812dab17-195f-4292-80a0-3ce34b9f2312)