Update timeline to display times in user timezone. (#3804)
Previously the timeline widget would display all times in UTC but other
times in the GUI are displayed in the user's selected timezone. This PR
changes this to make the timeline widget display times in the same
timezone selected by the user.

Also refactored CSV parameter editor to not use editable tables. These
were not intuitive as the user had to click on the table to realize they
were editable. Also it caused the layout to jump around. This PR fixes
it in favor of fixed simple text areas.

Additionally this PR patches the timeline widget to ensure ctrl-wheel
zooms in steps of 1 - the default upstream is to jump in steps of 10
which is way too much