Table context menu allows annotation to timeline
This allows any row from any table to be annotated to a global
timeline. This is allowed even if there is no natural time column in
the row (the time is set to 1980 in that case).

This feature allows to selectively annotating noteworthy rows from any
artifact/client/flow/hunt into a central place for reporting and
analysis without first importing a time series.