Added additional checks for inflight flow status
The Inflight flow status check sends a message to the client about
flows that are currently in flight. This PR adds an additional check
to verify if the flow has already been resolved before sending this
check. This helps to avoid sending a status check for a flow that is
already resolved which will result in an error state for the flow,
even though it is already completed.

Also added a flag to disable this entire feature which might be
necessary for unusual configurations, while the feature is pretty new
and might introduce too much load on busy servers.

Also added toolbar to hunt client view.