Added typed parameters to all notebooks (#3925)
This allows all notebooks (collection, hunt, event and global) to have
typed parameters. For non global notebooks, the parameters of the
original collection are preserved so they are available inside the
notebook.

The new GUI allows to update these parameters at any time - this give
the benefit of:

1. Notebook cells can be written to depend on active parameters that the
user can manipulate.
2. Parameters are now discoverable with the GUI easily showing which
additional parameters are available to the VQL inside the cell.