Expose the debug server to all admin users (#3934)
Previously the debug server had to be manually enabled and served on a
different port. This makes it difficult to use as one needs to change
the config and restart the server.

However, the debug server does not pose more risk than running the
profile plugins (it is basically the same thing) so it should be
permanently exposed to those users. This should make it easier to access
because it will always be available if needed.