Bugfix: Actively reduce memory use in memcache filestore (#3935)
This PR forces synchronous writing when memory utilization is exceeded
in the memcache filestore. By writing synchronously we can push back
against writers to ensure that memory can be cleared before we take on
new writes. Previously this flushing was asynchronous which means there
was no guarantee that memory was properly flushed - leading to an
uncontrolled increase in memory use.

Fixes: #3906