Propagate client config to queries running in the client.
This propagates things like prevent_execve into the query scope to
ensure it is inspected at all levels.

Also implemented a plugin deny list as a more robust way to prevent
some plugins and functions from running on the endpoint. The extends
the idea of prevent_execve to other potentially dangerous
plugins (e.g. rm). It is implemented as a deny list in addition to the
previous allow_list to make it easier to use.