Added Sigma Default Field Mappings (#4049)
Normally the Sigma Model needs to specify the field mappings into the
event. This makes it difficult to write rules targeting fields which are
present in the event but have no mapping yet. It also means most
mappings are fairly obvious.

For example consider the event:

{
...
    "EventData": {
        "ImageFileName": "WmiPrvSE.exe"
    }
}

To refer to the ImageFileName we would need a field mapping like:

ImageFileName =  x=>x.EventData.ImageFileName

However in a Sigma rule it is just as easy to specify:

detection:
   filename_condition:
      EventData.ImageFileName|contains: WmiPrvSE

This is both clear and unambigious.

This PR allows field names to have "." and refer directly to the event
itself.