Bugfix: Detect result set replay and drop it
In some situations it is possible for result sets to be
retransmitted (for example when there is a load balancer in
front). This PR detects when a result set upload is replayed and
ignores it. Without this detection, the result set will append a
repeated set of data which may corrupt the uploaded query results.