Rate limit client event updates on artifact modification. (#4162)
When artifacts are modified we need to recompile the client event tables
- this takes time. If the user modifies a lot of artifacts very quickly
(e.g. with a VQL query) this can overwhelm the server as it will try to
recompile the event table for each artifact.

This PR adds debouncing via the notification service to limit the rate
of event table compilation.