Bugfix: Do not complete collection on error.
The client will fail a collection when there is a log message at ERROR
level. Previously this tagged the collection as completed (with an
error). This means no further status was sent to the server. However,
the VQL continues running and sending back results.

This leaves the collection in an inconsistent state - the total count
of rows is incorrect, as more rows are still actually collected by the
client. The GUI may not show the correct results collected.

This PR does not consider an error to be represent the completion of
the collection - stat updates continue to be sent until the collection
is really finished.