Keep record of finished flows in client's flow manager. (#4202)
This allows profiles to see the status of recent flows that have
completed. It also addresses the issue of the in flight checks checking
the flow status after it has completed. The client will keep records of
past flows and simply return the exit status of those flows when asked.