Delay flow error until the flow is completed. (#4210)
Prevously the flow state was set to error as soon as a log was issued at
level ERROR however the query continued running until the end. This
caused the GUI to disable the cancel button so it was impossible to
cancel the flow.

This PR Keeps the flow at the IN_PROGRESS state until completion while
tagging the error message in the flow. This allows the flow to be
cancelled after the error by the GUI. After completion the flow will be
switched to the ERROR state.

Also:
1. Fix Registry accessor parsing of MULTI_SZ fields.
2. Fix Generic.Client.Profile artifact
3. Refactor tests to use the global mock clock