Implemented static analysis for client side permissions
Client side artifacts run on the endpoint without ACL
enforcement. This is usually what we want but sometimes the extra
permissions can give the user extra permissions on the end point.

The artifact writer can specify required permissions which will be
enforced by the server prior to collecting the artifact.

However, sometimes the artifact can use plugins with high priviledge
safely - in that case we do not want to restrict the users that may
collect it.

For example say the artifact collects autoruns by shelling to the
autoruns.exe tool. Even though it is using the execve() plugin this
call is safe because the args are fixed and can not be influenced by
the user. However if the artifact passed user input into the execve()
plugin, the user requires the EXECVE permission.

Previously, the artifact could declare EXECVE as a required
permission, for artifact uses where user input was directly allowed in
execve() calls. This enforces additional checks on the launching user
to ensure they have the EXECVE permission.

This PR modifies the artifact verifier to track plugin permissions
used in the artifact. This allows us to see if the artifact
inadvertantly gives the user permissions they do not have.

This PR introduces another field to the artifact definition called
`implied_permissions` where the artifact writer can declare
permissions which the artifact will give but the user does not require
those. This helps the verifier identify additional permissions that
are accidentally given to users on the client.

This PR adds a test to ensure built in artifacts have all necessary
permissions declared either in `required_permissions` or
`implied_permissions`