Refactored Linux Quarantine VQL and fixed tests
Fixed bug in mock plugin where it was impossible to provide different
mocks for separate calls. This was required for this test because
http_client() is used to call the server first, then the forbidden url
next so we need to have different mocks for first call and second call.