Refactored User Manager and ACL Manager to be in memory (#4290)
Previously there was a time based cache but we dont expect a lot of
users so we can reasonable keep the user list in memory permanenetly.
This helps to address potential problems in casing when saving user
accounts with different cases.

The new code deliberately catches these potential casing clashes and
ignores user accounts and acls for users who differ only in case. This
might help resolve #4271