Bugfix: Propagate column types from the artifact to notebook cells (#4318)
This fixes artifacts that defined column types - there were ignored in
notebooks. The `Format Column` GUI now also honors the column types and
shows them as preconfigured.

Also: Automatically create an upload preview column for shell artifacts
with large output. This makes it better to view in the GUI and possible
to handle a lot of output