Prevent access to parts of the filesystem (#4371)
On the server, VQL can access the entire filesystem, either with the
"file" accessor or the "fs" accessor. This allows a VQL query in the
notebook to read any internal files in the file store as well as other
sensitive files on the server

This PR restricts the file and fs accessors into a set of allowable,
preconfigured directories. Additionally VQL plugins that are allowed to
write to the filesystem also honor this restriction.

This PR also introduces a Data Encryption Key for secrets. This ensures
that even if the secret files are exposed the secrets themselves can not
be unlocked without access to the DEK.

The DEK can be specified in the config file or in an environment
variable.