Bugfix: The http_client() plugin did not honor regex filter. (#4379)
This PR fixes this issue as well as removing the ability to create or
modify secret templates. It has become apparent that user settable
secret templates are just not useful as the secret can never be exposed
to VQL so custom templates just can not be used by VQL at all.

Therefore this PR also refactores the secret API to remove this and just
initializes the hard coded templates in memory.

Also:

* Bugfix the column_filter() plugin to support queries with varying
columns per row.
* Disable editing of anything other than strings in the json_array GUI.