Remove unnecessary 'not /' characters from basePathRegEx (#4387)
The former regex in
https://github.com/Velocidex/velociraptor/blob/e16721ee09af6f8bfca40e14c7f5438d815c7013/services/sanity/frontend.go#L14
is misleading.

The 'not /' character `[^/]` at the beginning and end of the
basePathRegex forces you to have a base_path of at least 3 characters,
which does not match the error message.
Furthermore it also allows you to use not intended paths like:

/*a* - where * stands for every character not within the alphanumeric
character regex in between, which might cause problems or at least gives
the opportunity to build unfunctional urls.

The removal of those characters should be enough, as the regex not
implicitly does not allow `/` after the first character or at the end by
the nature of the ^$ signs.

@blobbfischer