Refactored accessors and improved permissions. (#4386)
The PR refactors the accessor API to add a descriptor for each accessor,
like vql plugins and functions. This allows a more structured approach
to documentation.

Additionally, the descriptor can specify the type of permissions
required to use the accessor, just like plugins. Some accessrs are
configured through scope variables, and this now adds a self documenting
args descriptor just like plugins.

This PR also introduces a new permission NETWORK. This permission
declares that the user is able to access the network and it is enforced
by various plugins/functions and accessors that access the network. This
additional permission provide finer grained control over user activity.