More work on nanny.
Added profile tracker for nanny and ring buffers.

Added fault injector to emulate stuck http connections - these seem to
be present sometimes after suspend.

Nanny has a mechanism now to warn clients they are out of compliance
and allow them an opportunity to rectify the issue. This is now used
in the send and receive HTTP connections to abort the current
connection and restart it.