Update collector remapping artifacts (#4560)
`Windows.Collectors.Remapping`
- Support containers secured with a fixed password (e.g. ones exported
from the server)
- Add precondition so that it doesn't produce a useless remapping when
the container doesn't contain uploads.
- Let the accessor to be consistently set from the parameters, since
there may be use cases for using other accessors such as
`collector_sparse` or `zip`.
- Log the hostname, for the record - allows users to find it in the GUI
without needing to go check the remapping file.
- More detailed description.

Note: I know that using ZIP_PASSWORDS as a parameter name is a bit ugly,
but it's necessary to allow the precondition to work at the top level.
Source-level preconditions are not viable because the test/verification
sources are dependent on the first source.

Improved the descriptions in the 2 KapeFiles artifacts. The name
references are currently a bit awkward as we transition to
`Windows.Triage.Targets`.

Changed GUI tooltips on 2 of the buttons since they didn't really match
their present functions.