fix: missing rpm release specification (#4587)
This pull request fixes the missing RPM release specification issue
observed in #4577.

It looks like when the RPM client package is created using the
`rpm_create` plugin, the release is not extracted from the template or
specification definition. Instead, the release is determined by the
`release` argument to the plugin. This happens in the
[expandSpec](https://github.com/Velocidex/velociraptor/blob/29442ef0525c87d1988343a680e4e3d5222a336e/vql/tools/packaging/package.go#L266)
function in `package.go` which directly takes in the `arg.Release`
argument.

To fix this, I have added a default `Release` ("A") to the RPM
specification and also modified the `Server.Utils.CreateLinuxPackages`
artifact to include a new parameter allowing users to specify a custom
release, as well as passing the parameter or default from the
specification to the `rpm_create` plugin.

I couldn't locate any tests for the `Server.Utils.CreateLinuxPackages`
artifact, however please let me know if I've missed them and I'll
update/add the tests for this change.