Bugfix: Filestore path types were inferred from the file extension (#4592)
However in certain paths in the filestore (e.g. file uploads in
collections) the files are always stored with
PATH_TYPE_FILESTORE_ANY. This meant that uploaded files with the same
extension as internal files were mistaken for different filestore paths.

This PR infers the pathspec type from the components list. It carves out
directories in the filestore which are untyped. All files stored there
will be assigned a PATH_TYPE_FILESTORE_ANY type, regardless of file
extension.